test fixes
- also consolidating all package tests into trip_tests.go
- this seems to work better with go tooling, because I can run each test in isolation this way...
